+++ Modules/Intent.Modules.DomainEvents/Templates/DomainEventServiceInterface/DomainEventServiceInterfaceTemplatePartial.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.DomainEvents.Templates.DomainEventBase;
@@ -14,7 +15,7 @@ using Intent.Templates;
 namespace Intent.Modules.DomainEvents.Templates.DomainEventServiceInterface
 {
     [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
-    partial class DomainEventServiceInterfaceTemplate : CSharpTemplateBase<object>
+    public partial class DomainEventServiceInterfaceTemplate : CSharpTemplateBase<object>, ICSharpFileBuilderTemplate
     {
         [IntentManaged(Mode.Fully)]
         public const string TemplateId = "Intent.DomainEvents.DomainEventServiceInterface";
@@ -22,7 +23,23 @@ namespace Intent.Modules.DomainEvents.Templates.DomainEventServiceInterface
         [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
         public DomainEventServiceInterfaceTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)
         {
+            CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
+                .AddUsing("System.Threading")
+                .AddUsing("System.Threading.Tasks")
+                .AddInterface($"IDomainEventService", @interface =>
+                {
+                    @interface.AddMethod("Task", "Publish", method =>
+                    {
+                        method
+                            .AddParameter(GetDomainEventBaseType(), "domainEvent")
+                            .AddParameter("CancellationToken", "cancellationToken", p => p.WithDefaultValue("default"))
+                            ;
+                    });
+                });
         }
+
+        [IntentManaged(Mode.Fully)]
+        public CSharpFile CSharpFile { get; }
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         protected override CSharpFileConfig DefineFileConfig()
@@ -30,6 +47,12 @@ namespace Intent.Modules.DomainEvents.Templates.DomainEventServiceInterface
             return new CSharpFileConfig(
                 className: $"IDomainEventService",
                 @namespace: $"{OutputTarget.GetNamespace()}");
+        }
+
+        [IntentManaged(Mode.Fully)]
+        public override string TransformText()
+        {
+            return CSharpFile.ToString();
         }
 
         private string GetDomainEventBaseType()