+++ Modules/Intent.Modules.SharedKernel.Consumer/FactoryExtensions/DependencyInjectionExtension.cs
@@ -1,0 +1,46 @@
+using Intent.Engine;
+using Intent.Modules.Common;
+using Intent.Modules.Common.Plugins;
+using Intent.Plugins.FactoryExtensions;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.FactoryExtension", Version = "1.0")]
+
+namespace Intent.Modules.SharedKernel.Consumer.FactoryExtensions
+{
+    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    public class DependencyInjectionExtension : FactoryExtensionBase
+    {
+        public override string Id => "Intent.SharedKernel.Consumer.DependencyInjectionExtension";
+
+        [IntentManaged(Mode.Ignore)]
+        public override int Order => 0;
+
+        /// <summary>
+        /// This is an example override which would extend the
+        /// <see cref="ExecutionLifeCycleSteps.AfterTemplateRegistrations"/> phase of the Software Factory execution.
+        /// See <see cref="FactoryExtensionBase"/> for all available overrides.
+        /// </summary>
+        /// <remarks>
+        /// It is safe to update or delete this method.
+        /// </remarks>
+        protected override void OnAfterTemplateRegistrations(IApplication application)
+        {
+            // Your custom logic here.
+        }
+
+        /// <summary>
+        /// This is an example override which would extend the
+        /// <see cref="ExecutionLifeCycleSteps.BeforeTemplateExecution"/> phase of the Software Factory execution.
+        /// See <see cref="FactoryExtensionBase"/> for all available overrides.
+        /// </summary>
+        /// <remarks>
+        /// It is safe to update or delete this method.
+        /// </remarks>
+        protected override void OnBeforeTemplateExecution(IApplication application)
+        {
+            // Your custom logic here.
+        }
+    }
+}