+++ Modules/Intent.Modules.MediatR.DomainEvents/release-notes.md
@@ -1,7 +1,7 @@
 ### Version 5.0.17
 
 - Improvement: `Primary Key`s with Data source set to `User supplied` or `Auto-generated` no longer explicItly call `UnitOfWork` save changes. `Default` still does and should be used if you have Database generated primary keys.
-
+- Improvement: Updated `DomainEventServiceTemplate` template to `CSharpFileBuilderTemplate`.
 
 ### Version 5.0.16